[Feature][Connector] Add kafka connector.

+++ chunjun-connectors/chunjun-connector-cassandra/src/main/java/com/dtstack/chunjun/connector/cassandra/converter/CassandraColumnConverter.java
@@ -28,6 +28,8 @@ import com.dtstack.chunjun.element.column.BigDecimalColumn;
 import com.dtstack.chunjun.element.column.BooleanColumn;
 import com.dtstack.chunjun.element.column.ByteColumn;
 import com.dtstack.chunjun.element.column.BytesColumn;
+import com.dtstack.chunjun.element.column.DoubleColumn;
+import com.dtstack.chunjun.element.column.FloatColumn;
 import com.dtstack.chunjun.element.column.StringColumn;
 import com.dtstack.chunjun.element.column.TimestampColumn;
 import com.dtstack.chunjun.throwable.UnsupportedTypeException;
@@ -121,9 +123,9 @@ public class CassandraColumnConverter
             case "INT":
                 return val -> new BigDecimalColumn((Integer) val);
             case "FLOAT":
-                return val -> new BigDecimalColumn((Float) val);
+                return val -> new FloatColumn((Float) val);
             case "DOUBLE":
-                return val -> new BigDecimalColumn((Double) val);
+                return val -> new DoubleColumn((Double) val);
             case "LONG":
             case "BIGINT":
                 return val -> new BigDecimalColumn((Long) val);
