+++ chunjun-connectors/chunjun-connector-sybase/src/main/java/com/dtstack/chunjun/connector/sybase/converter/SybaseSyncConverter.java
@@ -25,6 +25,8 @@ import com.dtstack.chunjun.element.AbstractBaseColumn;
 import com.dtstack.chunjun.element.column.BigDecimalColumn;
 import com.dtstack.chunjun.element.column.BooleanColumn;
 import com.dtstack.chunjun.element.column.BytesColumn;
+import com.dtstack.chunjun.element.column.DoubleColumn;
+import com.dtstack.chunjun.element.column.FloatColumn;
 import com.dtstack.chunjun.element.column.SqlDateColumn;
 import com.dtstack.chunjun.element.column.StringColumn;
 import com.dtstack.chunjun.element.column.TimeColumn;
@@ -85,12 +87,12 @@ public class SybaseSyncConverter extends JdbcSyncConverter {
                 return val -> {
                     if (val instanceof Double) {
                         BigDecimal b = new BigDecimal(String.valueOf(val));
-                        return new BigDecimalColumn(b.doubleValue());
+                        return new DoubleColumn(b.doubleValue());
                     }
-                    return new BigDecimalColumn((Float) val);
+                    return new FloatColumn((Float) val);
                 };
             case DOUBLE:
-                return val -> new BigDecimalColumn((Double) val);
+                return val -> new DoubleColumn((Double) val);
             case BIGINT:
                 return val -> {
                     if (val instanceof Integer) {
